refactor: consolidate app.py into main.py
