--- core/__init__.py
@@ -1,1 +1,0 @@
-"""Core package for the Lunch application."""
